t3php-mode: Removed "quit" functionality in toc buffer
I decided to remove the "quit" functionality in the toc buffer,
i.e. you can't delete the toc window and keep the toc buffer in
the background anymore. The buffer is always killed when it is
left instead. Thus you are guaranteed to get the up to date class
outline when it is requested.
With respect to clean code strategies a php class should not have
as many methods, that a modern computer lacks the perfomance of
parsing these methods every time you open the toc buffer.